+++ src/components/MyKiva/LendingStats.vue
@@ -249,6 +249,10 @@ export default {
 			type: Boolean,
 			default: false
 		},
+		goalsEntrypointEnable: {
+			type: Boolean,
+			default: false
+		}
 	},
 	data() {
 		return {